Refactor how Cart renders as a sidebar
I moved the cart sidebar to the header so it's the same component
through the different views.

But to reuse the same component for the /cart path, I had to change
the implementation to use props instead of useOutletContext() to
get the data it needs.

That caused changes in Home and Shop too, since those pages use the
Cart sidebar.